requirements.txt
cryptography==41.0.4
pyjwt==2.8.0
argon2-cffi==25.1.0
pytest
pytest-cov
